in order to identify the problems caused by integer type promotion when comparing unsigned and signed integers, in this series, we

- address the warnings raised by `-Wsign-compare` compiler option
- add `-Wsign-compare` compiler option to the building systems

Closes #14652

* github.com:scylladb/scylladb:
  treewide: use unsigned variable to compare with unsigned
  treewide: compare signed and unsigned using std::cmp_*()

+++ test/boost/chunked_managed_vector_test.cc
@@ -274,7 +274,7 @@ SEASTAR_TEST_CASE(test_correctness_when_crossing_chunk_boundary) {
         as(region, [&] {
             size_t max_chunk_size = lsa::chunked_managed_vector<int>::max_chunk_capacity();
 
-            lsa::chunked_managed_vector<int> v;
+            lsa::chunked_managed_vector<size_t> v;
             for (size_t i = 0; i < (max_chunk_size + 1); i++) {
                 v.push_back(i);
             }
